refactor(#163): blog newsletter pages - refactored pages to reduce duplication
