+++ components/Post/Pages/index.ts
@@ -1,0 +1,2 @@
+export { default as PostPage } from './PostPage';
+export { default as PostGridPage } from './PostGridPage';
